キノコタケノコ第一試験ノイズ研究所への改名とそれに伴う様々な変更
