+++ components/layout/About.tsx
@@ -1,0 +1,82 @@
+import Image from "next/image"
+
+import { Kiwi_Maru } from "next/font/google"
+
+const maruFont = Kiwi_Maru({
+  weight: "400",
+  subsets: ["latin"],
+})
+
+export default function About() {
+  return (
+    <article className="my-6 relative w-full px-6 max-w-2xl md:max-w-3xl lg:max-w-4xl xl:max-w-5xl mx-auto">
+      <div className="grid grid-cols-5 grid-rows-[1fr_2fr] gap-x-2 lg:gap-x-4 xl:gap-x-8 pb-5">
+        <div className="col-start-1 col-end-3 row-start-1 row-end-2">
+          <p className="text-[calc(0.8rem_+_5vw)] whitespace-nowrap">Now Season</p>
+        </div>
+
+        <div className="col-start-1 col-end-2 row-start-2 row-end-3 relative">
+          <Image src="/haru.svg" fill objectFit="contain" alt="春" />
+        </div>
+
+        <div className="col-start-2 col-end-3 row-start-2 row-end-3 relative">
+          <Image src="/pc.png"
+            fill
+            style={{ objectFit: "contain", imageRendering: "pixelated" }}
+            unoptimized
+            alt="PC"
+          />
+        </div>
+      </div>
+
+      <div className="prose-sm md:prose-md lg:prose-lg">
+        <h1 className={maruFont.className}>電子ライ麦畑はイーロンマスクを欺く</h1>
+        <h2 className={maruFont.className}>インターネットはただの通信網ではない。</h2>
+
+        <p>ようこそ、キノコタケノコ第一試験ノイズ研究所へ。人工電子戦争、すなわちキノコタケノコ戦争を、今ここから始めましょう。</p>
+        <p>私たちは、インターネットの片隅で「しょうもないこと」を真剣に考える人々のための畑を耕しています。</p>
+
+        <p>世界は常に「意味」を求めます。役に立つこと、効率的なこと、商業的価値のあるもの――</p>
+        <p>「意味」はストーリーを生み、人はそのストーリーを伝え合います。たとえば、「AがBにXをした。だからAはX&apos;なやつだ。」</p>
+        <p>しかし、伝えられた側は、AがBにXをした事実ではなく、AがX&apos;であるという「意味」を盲信してしまうのです。</p>
+
+        <p>SNSを埋め尽くすテキストの洪水は、この「意味」に焦点を当てています。SNSで怒る人々は、実際の行動ではなく、その背後にある「意味」に対して激昂しています。</p>
+        <p>「意味」はわかりやすい充実感を与え、崇高な行動のように見せかけますが、これこそが手段が形骸化し目的化する、本末転倒のメカニズムなのです。</p>
+
+        <p>イーロン・マスクがツイッターをPvPに改変したのは、「しょうもない独り言」を駆逐し「意味のあるコンテンツ」を創出させるためのアルゴリズム改造に他なりませんでした。</p>
+        <p>しかし、PvP化したツイッターは、過激でうんざりする体験ばかりを生み出す結果となりました。</p>
+        <p>サブカルひねくれ野郎のお家芸、ゼロ年代の讃歌を歌うわけではありませんが、本当に面白かったのは、意味のないノイズが溢れていたあの時代だったのではないでしょうか？</p>
+
+        <p>「きのこの山たけのこのさと」は、そんな失われた「ノイズ」を再び咲かせるための場所です。</p>
+        <p>そして、あなたの「しょうもない」が、この畑の土壌を豊かにしてくれるのです。</p>
+
+        <h2 className={maruFont.className}>キノコタケノコ第一試験ノイズ研究所の使命</h2>
+        
+        <p>私たちの使命は、イーロン・マスクのアルゴリズムを欺くためのノイズを開発することにあります。</p>
+        <p>彼のアルゴリズムは過激なレスバトルを求める傾向にありますが、我々は「キノコタケノコ論争」と呼ばれるネットミームを活用します。</p>
+        <p>ディスコード内であらかじめ策定されたシナリオに基づき、「きのこの山」と「たけのこの里」による壮大なレスバトルを演じ、ツイッターのアルゴリズムを欺いて多数のユーザーのタイムラインに現れることを狙います。</p>
+        <p>我々の計画的な「キノコタケノコ論争」は、イーロン・マスクが望む「意味のあるコンテンツ」とは一線を画し、ユーザーのタイムラインは、計画的に運用された意味のないレスバトルで埋め尽くされ、インターネットにおける「意味とアルゴリズム」の崩壊を招くのです。</p>
+
+
+        <h2 className={maruFont.className}>インターネットの原点と未来</h2>
+        
+        <p>インターネットは、もともとカオスの温床でした。</p>
+        <p>意味不明な掲示板の書き込み、用途不明のフリーソフト、誰にも見られない個人ホームページ――</p>
+        <p>あの混沌は、企業のアルゴリズムによって「除草」され、今や美しく舗装されたハイウェイのようになってしまいました。</p>
+        <p>しかし、私たちは知っています。</p>
+        <p>道端に咲く、意味のない花こそが、一番美しいと。</p>
+
+        <p>我々は、企業によって作られた資本主義に基づく「意味」にサヨナラを告げる必要があります。魂と魂が結びつく回路こそが、インターネット本来の唯一の「意味」だったのです。</p>
+        <p>きれいに舗装されたハイウェイは、広告収入とコンテンツ最適化のために設計された、人間性を犠牲にするツイッターという廃墟のような電子空間を維持するに過ぎません。</p>
+        <p>まるでゼロ年代のセカイ系ヒロインのように私達はツイッターというセカイの維持のために人柱となっているのです。</p>
+        <p>かつて、Lainが描いたように、人々の意識がインターネットという電子回路に集合し、一つの神経回路、巨大な脳、そして壮大な意識を形成していた世界は、今や失われてしまいました。</p>
+
+        <h2 className={maruFont.className}>我々のビジョン</h2>
+        <p>我々は、「キノコタケノコ論争」による計画的な人工ノイズの運用を通じて、アルゴリズムを欺き、コンテンツ型SNSの内部崩壊を引き起こし、</p>
+        <p>再び人と人が繋がるコミュニケーションとしてのSNSを復興させることを目指します。</p>
+        <p>そのためには、アルゴリズムを欺く優れた「ノイズ」の研究開発と、作戦会議、そして実際に「ノイズ」を運用するあなたたちの協力が不可欠です。</p>
+        <p>人間的な混沌を再びネットに取り戻すため、ぜひ我々キノコタケノコ第一試験ノイズ研究所のディスコードに参加し、共に新たな電子戦争を繰り広げましょう！</p>
+      </div>
+    </article>
+  )
+}
