+++ tailwind.config.ts
@@ -2,6 +2,7 @@ import type { Config } from "tailwindcss";
 import plugin from 'tailwindcss/plugin';
 
 export default {
+  important: true,
   content: [
     "./pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./components/**/*.{js,ts,jsx,tsx,mdx}",
